refactor: moved add-cam-to-inventory function to AddGear (#20)
