+++ src/entrypoints/popup/output.css
@@ -458,10 +458,6 @@ video {
 .pt-2 {
     padding-top: 0.5rem;
 }
-.text-2xl {
-    font-size: 1.5rem;
-    line-height: 2rem;
-}
 .text-lg {
     font-size: 1.125rem;
     line-height: 1.75rem;
@@ -469,6 +465,10 @@ video {
 .text-sm {
     font-size: 0.875rem;
     line-height: 1.25rem;
+}
+.text-xl {
+    font-size: 1.25rem;
+    line-height: 1.75rem;
 }
 .font-bold {
     font-weight: 700;
